add default values

+++ components/huawei_r4850/number/huawei_r4850_number.h
@@ -9,14 +9,14 @@ namespace huawei_r4850 {
 
 class HuaweiR4850Number : public number::Number, public Component {
 public:
-  void set_parent(HuaweiR4850Component *parent, int8_t functionCode) {
+  void set_parent(HuaweiR4850Component *parent, uint16_t functionCode) {
     this->parent_ = parent;
     this->functionCode_ = functionCode;
   };
 
 protected:
   HuaweiR4850Component *parent_;
-  int8_t functionCode_;
+  uint16_t functionCode_;
 
   void control(float value) override;
 };
